+++ mysql/tables/llx_commande_fournisseurdet.sql
@@ -33,9 +33,9 @@ create table llx_commande_fournisseurdet
   remise_percent             real         DEFAULT 0,    -- pourcentage de remise
   remise                     real         DEFAULT 0,    -- montant de la remise
   price                      real,                      -- prix final
-  subprice                   double(16,8) DEFAULT 0,    -- prix unitaire
-  total_ht                   double(16,8) DEFAULT 0,    -- Total HT de la ligne toute quantité et incluant remise ligne et globale
-  total_tva                  double(16,8) DEFAULT 0,	  -- Total TVA de la ligne toute quantité et incluant remise ligne et globale
-  total_ttc                  double(16,8) DEFAULT 0,	  -- Total TTC de la ligne toute quantité et incluant remise ligne et globale
+  subprice                   double(24,8) DEFAULT 0,    -- prix unitaire
+  total_ht                   double(24,8) DEFAULT 0,    -- Total HT de la ligne toute quantité et incluant remise ligne et globale
+  total_tva                  double(24,8) DEFAULT 0,	  -- Total TVA de la ligne toute quantité et incluant remise ligne et globale
+  total_ttc                  double(24,8) DEFAULT 0,	  -- Total TTC de la ligne toute quantité et incluant remise ligne et globale
   info_bits	                 integer      DEFAULT 0     -- TVA NPR ou non
 )type=innodb;
