+++ mysql/tables/llx_contratdet.sql
@@ -43,12 +43,12 @@ create table llx_contratdet
   tva_tx                double(6,3)   DEFAULT 0, 	         -- taux tva
   qty                   real          NOT NULL,            -- quantité
   remise_percent        real          DEFAULT 0,    		   -- pourcentage de remise
-  subprice              double(16,8)  DEFAULT 0,           -- prix unitaire
+  subprice              double(24,8)  DEFAULT 0,           -- prix unitaire
   price_ht              real,              		             -- prix final (obsolete)
   remise                real          DEFAULT 0,    		             -- montant de la remise (obsolete)
-  total_ht              double(16,8)  DEFAULT 0,     		   -- Total HT de la ligne toute quantité et incluant remise ligne et globale
-  total_tva             double(16,8)  DEFAULT 0,	   		   -- Total TVA de la ligne toute quantité et incluant remise ligne et globale
-  total_ttc             double(16,8)  DEFAULT 0,	   		   -- Total TTC de la ligne toute quantité et incluant remise ligne et globale
+  total_ht              double(24,8)  DEFAULT 0,     		   -- Total HT de la ligne toute quantité et incluant remise ligne et globale
+  total_tva             double(24,8)  DEFAULT 0,	   		   -- Total TVA de la ligne toute quantité et incluant remise ligne et globale
+  total_ttc             double(24,8)  DEFAULT 0,	   		   -- Total TTC de la ligne toute quantité et incluant remise ligne et globale
   info_bits		          integer       DEFAULT 0, 		       -- TVA NPR ou non
 
   fk_user_author        integer       NOT NULL DEFAULT 0,
