+++ mysql/tables/llx_product_fournisseur_price_log.sql
@@ -25,7 +25,7 @@ create table llx_product_fournisseur_price_log
   rowid                       integer AUTO_INCREMENT PRIMARY KEY,
   datec                       datetime,
   fk_product_fournisseur      integer      NOT NULL,
-  price                       double(16,8) DEFAULT 0,
+  price                       double(24,8) DEFAULT 0,
   quantity                    double,
   fk_user                     integer
