+++ mysql/tables/llx_societe_remise_except.sql
@@ -27,9 +27,9 @@ create table llx_societe_remise_except
   rowid           integer AUTO_INCREMENT PRIMARY KEY,
   fk_soc          integer NOT NULL, -- client
   datec           datetime,
-  amount_ht       double(16,8) NOT NULL,
-  amount_tva 	  double(16,8) DEFAULT 0 NOT NULL,
-  amount_ttc 	  double(16,8) DEFAULT 0 NOT NULL,
+  amount_ht       double(24,8) NOT NULL,
+  amount_tva 	  double(24,8) DEFAULT 0 NOT NULL,
+  amount_ttc 	  double(24,8) DEFAULT 0 NOT NULL,
   tva_tx   	      double(6,3)  DEFAULT 0 NOT NULL,
   fk_user           integer NOT NULL,
   fk_facture        integer,
